# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Bookworm API
  version: 0.1.0
  description: Mandatory paths/schemas per MASTER-BUILD-SPEC section 9. All endpoints versioned under /v1.
servers:
  - url: http://localhost:3001
...
paths:
...
  /v1/chapters/{chapterId}/operations:
    post:
      tags: [v1]
      operationId: applyChapterOperation
      summary: Apply document operation (optimistic concurrency via expectedVersion)
      parameters:
        - name: chapterId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentOperation" }
      responses:
        "200":
          description: New document version
          content:
            application/json:
              schema:
                type: object
                required: [version]
                properties:
                  version: { type: integer }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DocumentOperation:
      type: object
      required: [operationId, type, target, payload, expectedVersion]
      properties:
        operationId: { type: string }
        type: { type: string }
        target: { type: object }
        payload: { type: object }
        source: { type: string, enum: [human, ai] }
        sourceRef: { type: string }
        expectedVersion: { type: integer }
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object }
...
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
security:
  - bearerAuth: []